OO solution
