+++ src/job/ZosJobsProvider.ts
@@ -17,7 +17,7 @@ import { IProfileLoaded, Logger, IProfile, ISession } from "@zowe/imperative";
 import { Profiles, ValidProfileEnum } from "../Profiles";
 import { Job } from "./ZoweJobNode";
 import { getAppName, sortTreeItems, labelHack } from "../shared/utils";
-import { FilterItem, FilterDescriptor, resolveQuickPickHelper, errorHandling, setProfile, setSession, } from "../utils";
+import { FilterItem, FilterDescriptor, resolveQuickPickHelper, errorHandling } from "../utils";
 import { IZoweTree } from "../api/IZoweTree";
 import { IZoweJobTreeNode } from "../api/IZoweTreeNode";
 import { ZoweTreeProvider } from "../abstract/ZoweTreeProvider";
@@ -431,18 +431,6 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
         }
     }
 
-    public async editSession(node: IZoweJobTreeNode) {
-        const profile = node.getProfile();
-        const profilename = node.getProfileName();
-        const EditSession = await Profiles.getInstance().editSession(profile, profilename);
-
-        if (EditSession) {
-            node.getProfile().profile= EditSession as IProfile;
-            setProfile(node, EditSession as IProfile);
-            setSession(node, EditSession as ISession);
-            this.refresh();
-        }
-    }
 
     public async onDidChangeConfiguration(e: vscode.ConfigurationChangeEvent) {
         if (e.affectsConfiguration(ZosJobsProvider.persistenceSchema)) {
